+++ CHANGELOG.md
@@ -12,9 +12,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Jersey to version 3.1.10 [#340](https://github.com/IN-CORE/incore-services/issues/340)
 - Swagger to version 2.2.28 [#251](https://github.com/IN-CORE/incore-services/issues/251)
 - Endpoint to delete and edit layers in visualization [#360](https://github.com/IN-CORE/incore-services/issues/360)
+
 ### Added
 - More fields in the visualziation layer model [#365](https://github.com/IN-CORE/incore-services/issues/365)
 - Added PATCH endpoint to update visualization fields, including layer order and layers [#373](https://github.com/IN-CORE/incore-services/issues/373)
+- Project tool endpoint that initiate NSI data creation and registration [#377](https://github.com/IN-CORE/incore-services/issues/377)
 - Endpoint to generate NSI based building inventory dataset by fips code [#370](https://github.com/IN-CORE/incore-services/issues/370)
 - Endpoint to download the generated NSI based building inventory dataset by fips code [#377](https://github.com/IN-CORE/incore-services/issues/377)